tickets display comm
