+++ src/styled/Styles.js
@@ -1,3 +1,4 @@
+
 import styled from 'styled-components';
 
 export const Wrapper = styled.div`
