--- src/styled/Ticket/Styles.js
@@ -1,22 +1,0 @@
-import styled from 'styled-components';
-
-export const Styles = styled.div`
-  .card {
-    width: 500px;
-    min-height: 185px;
-    padding: 20px 50px 25px 20px;
-    margin-top: 20px;
-    margin-right: 100px;
-  },
-  .card-body {
-    width: 500px;
-    margin: 0;
-    padding: 0;
-  }
-`;
-
-export const Title = styled.div`
-  .card-title {
-    color: #2196F3;
-  }
-`;
